requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
django-celery-beat==2.5.0
Pillow==10.1.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
whitenoise==6.6.0
gunicorn==21.2.0